Refactor utils module and implement signature generation and verification
