+++ Cargo.toml
@@ -19,6 +19,7 @@ tracing-subscriber = "0.3.17"
 sha2 = "0.10"
 sha3 = "0.10"
 once_cell = "1.19.0"
+ethers = "2.0.4"
 
 skde = { git = "https://github.com/radiusxyz/skde", rev = "d6ef27d4f1e34b391eddb039cf0b9f994f7cc2fe" }
 radius-sdk = { git = "https://github.com/radiusxyz/radius-sdk-rs", tag = "v0.1", features = [
