+++ data/authority/skde_params.json
@@ -6,5 +6,5 @@
     "h": "4294967296",
     "max_sequencer_number": "2"
   },
-  "signature": "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+  "signature": "0xe6c88ce4ba6d1dd9a9de4db124fe1710adfa3712ceeb80c4369b76d5bf44260e70b6916ff25f32ebb5f65cf4b0be90974ea226f0529f9a8d38cd365c14ff5cc71b"
 }
