+++ src/lib.rs
@@ -8,7 +8,7 @@ pub mod utils;
 pub use error::Error;
 pub use state::AppState;
 pub use types::*;
-pub use utils::get_current_timestamp;
+pub use utils::time::get_current_timestamp;
 
 #[cfg(test)]
 pub mod tests;
