+++ src/rpc/authority/get_authorized_skde_params.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use skde::delay_encryption::SkdeParams;
 
-use crate::{rpc::prelude::*, task::authority_setup::SignedSkdeParams, utils::create_signature};
+use crate::{rpc::prelude::*, task::authority_setup::SignedSkdeParams, utils::signature::create_signature};
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct GetAuthorizedSkdeParams;
@@ -21,7 +21,7 @@ impl RpcParameter<AppState> for GetAuthorizedSkdeParams {
     async fn handler(self, context: AppState) -> Result<Self::Response, RpcError> {
         let skde_params = context.skde_params().clone();
 
-        let signature = create_signature(&skde_params);
+        let signature = create_signature(&context, &skde_params).unwrap();
 
         let signed_skde_params = SignedSkdeParams {
             params: skde_params,
@@ -45,22 +45,3 @@ impl GetAuthorizedSkdeParamsResponse {
         self.signed_skde_params.params
     }
 }
-// impl From<SkdeParams> for GetAuthorizedSkdeParamsResponse {
-//     fn from(params: SkdeParams) -> Self {
-//         Self {
-//             skde_params: params,
-//         }
-//     }
-// }
-
-// impl GetAuthorizedSkdeParamsResponse {
-//     pub fn into_skde_params(self) -> SkdeParams {
-//         SkdeParams {
-//             t: self.skde_params.t,
-//             n: self.skde_params.n,
-//             g: self.skde_params.g,
-//             h: self.skde_params.h,
-//             max_sequencer_number: self.skde_params.max_sequencer_number,
-//         }
-//     }
-// }
