+++ src/task/authority_setup.rs
@@ -1,11 +1,11 @@
 use std::{fs, path::PathBuf};
 
-use radius_sdk::signature::Signature;
+use radius_sdk::signature::{ChainType, PrivateKeySigner, Signature};
 use serde::{Deserialize, Serialize};
 use skde::delay_encryption::{setup, SkdeParams};
 use tracing::{info, warn};
 
-use crate::{utils::create_signature, ConfigPath};
+use crate::ConfigPath;
 
 // Constants for SKDE setup parameters
 // TODO: Getting constants in a form of json file(?)
@@ -23,19 +23,36 @@ pub struct SignedSkdeParams {
 pub fn run_setup_skde_params(path: &ConfigPath) {
     let config_dir: PathBuf = path.as_ref().into();
     let skde_path = config_dir.join("skde_params.json");
+    let signing_key_path = config_dir.join("signing_key");
 
     if skde_path.exists() {
         warn!("SKDE parameter file already exists: {:?}", skde_path);
         return;
     }
 
-    let t = DEFAULT_TIME_PARAM_T;
-    let g = DEFAULT_GENERATOR.into();
-    let max = DEFAULT_MAX_SEQUENCER_NUMBER.into();
+    let signing_key_hex = match fs::read_to_string(&signing_key_path) {
+        Ok(key) => key.trim().to_string(),
+        Err(e) => {
+            warn!("Failed to read signing key from {:?}: {}", signing_key_path, e);
+            return;
+        }
+    };
 
-    let params = setup(t, g, max);
+    let signer = match PrivateKeySigner::from_str(ChainType::Ethereum, &signing_key_hex) {
+        Ok(signer) => signer,
+        Err(e) => {
+            warn!("Failed to create signer from signing key: {:?}", e);
+            return;
+        }
+    };
 
-    let signature = create_signature(&params);
+    let params = setup(
+        DEFAULT_TIME_PARAM_T,
+        DEFAULT_GENERATOR.into(),
+        DEFAULT_MAX_SEQUENCER_NUMBER.into(),
+    );
+
+    let signature = signer.sign_message(&params).unwrap();
 
     let signed_params = SignedSkdeParams { params, signature };
