+++ src/utils/time.rs
@@ -1,0 +1,8 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
+pub fn get_current_timestamp() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs()
+}
